requirements.txt
streamlit
google-generativeai
pandas
numpy
scikit-learn
PyPDF2
python-docx
